+++ Radial/Radial.cpp
@@ -368,45 +368,45 @@ private:
                 double dy = (p.y - (_btmLeft.y + (_btmLeft.y + _size.y)) / 2) / (_size.y/2);
 
                 if (dx >= 1 || dy >= 1) {
-                    tmpPix[0] = _color0.r;
-                    tmpPix[1] = _color0.g;
-                    tmpPix[2] = _color0.b;
-                    tmpPix[3] = _color0.a;
+                    tmpPix[0] = (float)_color0.r;
+                    tmpPix[1] = (float)_color0.g;
+                    tmpPix[2] = (float)_color0.b;
+                    tmpPix[3] = (float)_color0.a;
                 } else {
                     double dsq = dx*dx + dy*dy;
 
                     if (dsq >= 1) {
-                        tmpPix[0] = _color0.r;
-                        tmpPix[1] = _color0.g;
-                        tmpPix[2] = _color0.b;
-                        tmpPix[3] = _color0.a;
+                        tmpPix[0] = (float)_color0.r;
+                        tmpPix[1] = (float)_color0.g;
+                        tmpPix[2] = (float)_color0.b;
+                        tmpPix[3] = (float)_color0.a;
                     } else if (dsq <= 0 || _softness == 0) {
-                        tmpPix[0] = _color1.r;
-                        tmpPix[1] = _color1.g;
-                        tmpPix[2] = _color1.b;
-                        tmpPix[3] = _color1.a;
+                        tmpPix[0] = (float)_color1.r;
+                        tmpPix[1] = (float)_color1.g;
+                        tmpPix[2] = (float)_color1.b;
+                        tmpPix[3] = (float)_color1.a;
                     } else {
-                        double t = (1. - std::sqrt(dsq)) / _softness;
+                        float t = (1.f - (float)std::sqrt(dsq)) / (float)_softness;
                         if (t >= 1) {
-                            tmpPix[0] = _color1.r;
-                            tmpPix[1] = _color1.g;
-                            tmpPix[2] = _color1.b;
-                            tmpPix[3] = _color1.a;
+                            tmpPix[0] = (float)_color1.r;
+                            tmpPix[1] = (float)_color1.g;
+                            tmpPix[2] = (float)_color1.b;
+                            tmpPix[3] = (float)_color1.a;
                         } else {
-                            t = rampSmooth(t);
+                            t = (float)rampSmooth(t);
 
                             if (_plinear) {
                                 // it seems to be the way Nuke does it... I could understand t*t, but why t*t*t?
                                 t = t*t*t;
                             }
-                            tmpPix[0] = _color0.r * (1 - t) + _color1.r * t;
-                            tmpPix[1] = _color0.g * (1 - t) + _color1.g * t;
-                            tmpPix[2] = _color0.b * (1 - t) + _color1.b * t;
-                            tmpPix[3] = _color0.a * (1 - t) + _color1.a * t;
+                            tmpPix[0] = (float)_color0.r * (1.f - t) + (float)_color1.r * t;
+                            tmpPix[1] = (float)_color0.g * (1.f - t) + (float)_color1.g * t;
+                            tmpPix[2] = (float)_color0.b * (1.f - t) + (float)_color1.b * t;
+                            tmpPix[3] = (float)_color0.a * (1.f - t) + (float)_color1.a * t;
                         }
                     }
                 }
-                double a = tmpPix[3];
+                float a = tmpPix[3];
 
                 // ofxsMaskMixPix takes non-normalized values
                 tmpPix[0] *= maxValue;
@@ -425,26 +425,26 @@ private:
                     }
                 }
                 if (processR) {
-                    tmpPix[0] = tmpPix[0] + srcPixRGBA[0]*(1.-a);
+                    tmpPix[0] = tmpPix[0] + srcPixRGBA[0]*(1.f-a);
                 } else {
                     tmpPix[0] = srcPixRGBA[0];
                 }
                 if (processG) {
-                    tmpPix[1] = tmpPix[1] + srcPixRGBA[1]*(1.-a);
+                    tmpPix[1] = tmpPix[1] + srcPixRGBA[1]*(1.f-a);
                 } else {
                     tmpPix[1] = srcPixRGBA[1];
                 }
                 if (processB) {
-                    tmpPix[2] = tmpPix[2] + srcPixRGBA[2]*(1.-a);
+                    tmpPix[2] = tmpPix[2] + srcPixRGBA[2]*(1.f-a);
                 } else {
                     tmpPix[2] = srcPixRGBA[2];
                 }
                 if (processA) {
-                    tmpPix[3] = tmpPix[3] + srcPixRGBA[3]*(1.-a);
+                    tmpPix[3] = tmpPix[3] + srcPixRGBA[3]*(1.f-a);
                 } else {
                     tmpPix[3] = srcPixRGBA[3];
                 }
-                ofxsMaskMixPix<PIX, nComponents, maxValue, true>(tmpPix, x, y, srcPix, _doMasking, _maskImg, _mix, _maskInvert, dstPix);
+                ofxsMaskMixPix<PIX, nComponents, maxValue, true>(tmpPix, x, y, srcPix, _doMasking, _maskImg, (float)_mix, _maskInvert, dstPix);
             }
         }
     }
